Mejores Practicas Se Realizaron Los controlladores de las respetivas vistas Se Mejoraron los Estilos Se corregieron varios Bugs
