+++ SistemaMerck.Modelos/ViewModels/UsuarioVM.cs
@@ -12,18 +12,19 @@ namespace SistemaMerck.Modelos.ViewModels
     public class UsuarioVM
     {
 
-        [Display(Name = "Edad Actual")]
-        [Required(ErrorMessage = "Por favor, selecciona la Edad Actual.")]
+        [Required(ErrorMessage = "La Edad Actual es obligatoria.")]
+        [Range(8, 50, ErrorMessage = "La Edad Actual debe estar entre 8 y 50.")]
         public int EdadActual { get; set; }
 
-        [Display(Name = "Edad Primera Menstruación")]
-        [Required(ErrorMessage = "Por favor, selecciona la Edad de la Primera Menstruación.")]
-        [Range(typeof(int), "8", "50", ErrorMessage = "La Edad Primera Menstruación no puede ser mayor que la Edad Actual.")]
+        [Required(ErrorMessage = "La Edad Primera Menstruación es obligatoria.")]
+        [Range(8, 15, ErrorMessage = "La Edad Primera Menstruación debe estar entre 8 y 15.")]
         public int EdadPrimeraMentruacion { get; set; }
 
         public double ReservaOvarica { get; set; }
 
-        public List<SelectListItem> Edades { get; set; }
+        public List<SelectListItem> EdadesActuales { get; set; }
+
+        public List<SelectListItem> EdadesPrimeraMentruacion { get; set; }
     }
 
 }
