+++ SistemaMerck/Views/Home/Bienvenida.cshtml
@@ -1,0 +1,17 @@
+﻿@{
+    ViewData["Title"] = "Bienvenida";
+    Layout = "~/Views/Shared/_Layout.cshtml";
+}
+
+<body style="background-color: #2EBDCD; color: white; text-align: center;">
+
+    <div class="container mt-5">
+        <img class="img-fluid" src="~/imagen/inicio.png" alt="Logo" />
+    </div>
+
+    <div class="container mt-4">
+        <a href="@Url.Action("Index", "Home")" class="btn btn-outline-light btn-lg btn-rounded custom-btn">Comenzar</a>
+    </div>
+
+</body>
+
