--- SistemaMerck/Views/Home/Pantalla2.cshtml
@@ -1,46 +1,0 @@
-﻿@model SistemaMerck.Modelos.ViewModels.UsuarioVM
-@{
-    ViewData["Title"] = "Pantalla2";
-    Layout = "~/Views/Shared/_Layout.cshtml";
-}
-
-<body style="background: linear-gradient(to bottom, #c33b80 50%, #fdddca 50%)">
-    <div class="container mt-5">
-        <div class="row">
-            <div class="col-md-6 offset-md-3 text-center">
-                <p class="reserva-texto text-white font-family-roboto">EDAD ACTUAL: @Model.EdadActual</p>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col-md-6 offset-md-3 text-center">
-                <p class="reserva-texto text-white font-family-roboto">1ERA MENSTRUACIÓN MENSTRUACIÓN: @Model.EdadPrimeraMentruacion</p>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col-md-6 offset-md-3 text-center">
-                <a asp-controller="Home" asp-action="Pantalla3" class="btn btn-primary">Quiero Saber más!</a>
-            </div>
-        </div>
-    </div>
-    <div id="container" class="container" style="height: 60vh; margin: 0 auto; margin-top: 160px;"></div>
-    <script src="https://code.highcharts.com/highcharts.js"></script>
-    <script src="https://cdn.jsdelivr.net/gh/highcharts/highcharts@v7.0.0/modules/data.js"></script>
-    <script src="https://cdn.jsdelivr.net/gh/highcharts/highcharts@v7.0.0/modules/drilldown.js"></script>
-    <script src="https://cdn.jsdelivr.net/gh/highcharts/highcharts@v7.0.0/modules/boost.js"></script>
-    <script src="https://cdn.jsdelivr.net/gh/highcharts/highcharts@v7.0.0/modules/exporting.js"></script>
-    <script src="https://cdn.jsdelivr.net/gh/highcharts/highcharts@v7.0.0/modules/export-data.js"></script>
-    <script src="https://cdn.jsdelivr.net/gh/highcharts/highcharts@v7.0.0/modules/accessibility.js"></script>
-    <script>
-        var reservaOvarica = @Html.Raw(Json.Serialize(Model.ReservaOvarica));
-    </script>
-
-    <!-- Incluir tu script personalizado -->
-    <script src="~/js/chartConfig.js"></script>
-</body>
-
-
-
-
-
-
-
