+++ SistemaMerck/wwwroot/css/site.css
@@ -27,3 +27,4 @@ body {
 }
 
 
+/*style="background-color: @(ViewData["NavbarColor"]); border-bottom: 1px solid #dee2e6; box-shadow: 0 0.125rem 0.25rem rgba(0, 0, 0, 0.075);*/
