--- SistemaMerck/wwwroot/js/formularioPartial.js
@@ -1,11 +1,0 @@
-﻿function mostrarFormulario() {
-    $.get('/Home/FormularioContacto', function (data) {
-        $('#formularioContainer').html(data);
-    });
-}
-
-$(document).ready(function () {
-    $('#mostrarFormulario').click(function () {
-        mostrarFormulario();
-    });
-});
